+++ test/extract/extract_test.exs
@@ -1,7 +1,7 @@
 defmodule Text.ExtractTest do
   use ExUnit.Case, async: true
   doctest Text.Extract
-  doctest Text.Extract.Boundary
+  doctest Text.Extract.Link
   doctest Text.Extract.Email
   doctest Text.Extract.Scanner
   doctest Text.Extract.Script